Add an npm script to regenerate Flow types (#62)
The file `flow_deps/flow.d.ts` contains Typescript type definitions for
all the Flow catalog types. This commit adds a slightly easier way of
updating those types to the latest version. This is achieved by adding
`fetch-flow.sh`, which downloads the most recent build of `flowctl`, and
then calling that in an npm script to re-generate the types.