Allow rendering JSON forms from wacky schemas
The EndpointConfigForm and ResourceConfigForm previously would barf
whenever they encountered a non-compliant JSON schema. This commit makes
the AJV validator much more permissive in the schemas that it will
accept.

This also removes the `json-refs` dependency and assumes that all json
schemas are fully self-contained. This property is ensured by the build
process for flow collection schemas. For endpoint and resource schemas,
it seems to also generally hold.