Add nginx.conf so that page refreshes will work
React-router wasn't playing nicely with the default nginx config.
This adds a config file, based on the default, which has a `try_files`
that falls back to index.html, which is what react-router needs.
With this, you should be able to refresh on any page and stay where
you're at in the UI.