Use local config-encryption service (#263)
When running the UI locally, use a local instance of the
config-encryption service. UI developers may still switch back to the
production config-encryption service when running without a data plane.
But normally, the local config-encryption service is needed, since
developers will not be able to decrypt credentials that were encrypted
using the production service.