Make open graph title optional for capture connectors
Previously, the open graph title was used exclusively as the display
name of a connector on the captures table page, which relies on the
assumption that open graph data will always be present for every capture
displayed in the UI. This assumption is not always valid, since we have
the ability to use custom images via the CLI. The result is that the
captures page would show as totally blank when any capture used a
connector that didn't have an open graph title.

This commit updates the capture rows to fallback to using the connector
image if the open graph title is not present. It also updates the query
to retrieve the image directly from the capture spec instead of the open
graph. The result is that the captures table can now render when a
capture uses a connector that isn't present in the database.