Add new Flowctl auth pages
Adds support for a new way for `flowctl` to authenticate.
First, the `/cli-auth/login` page was added. Once a user logs in there,
they are redirected to `/cli-auth/success`. Both of those pages
use separate Context providers so that they will use an entirely
separate go-true session and local storage entry than the rest of the
web app. Effectively, `flowctl` has its own separate session, and only
it will attempt to refresh the credentials for it.

As part of this, I had to do a bit of refactoring to allow re-using our
existing `Login` component with different redirect urls. I also fixed an
issue where redirects weren't working if the user manually entered the
OTP rather than clicking the link from the email. In that case, we need
to handle the redirect ourselves, since the `/auth/v1/validateToken`
endpoint will not respond with redirects.