Show task endpoints after successful publication
When a capture or materialization connector exposes one or more
endpoints, we now display a link to those endpoints in the
LogDialogActions component.

The `TaskEndpoint` component was added for rendering the endpoint info.
This component displays only a single endpoint per task, since it's
designed to fit into the small space at the bottom of the logs dialog.
The plan is for a later PR to introduce another component that
shows all of the endpoints for a task. But today, no connector exposes
more than one endpoint, so it's a non-issue.

Determining the endpoint URL requires knowledge of the data-plane
hostname. In order to get the hostname, this uses the `gateway_auth`
RPC to get an auth token that's scoped to the task. This is a bit
overkill, but was done with two things in mind. One is that there's
imminent plans to implement a data-plane authentication page that
generates this token and navigates the user to a data-plane-gateway
redirect handler, and generating these scoped auth tokens will be a
requirement for that. Second is that the control plane doesn't currently
offer any other endpoint for determining the data-plane URL for a task,
so re-using gateway_auth was just expedient.