Add data-plane authorization page
Flow tasks may expose private network ports, which can only be accessed
using a data-plane access token that authorizes the user to access the
task. This has been a difficult feature to use because we provided no
easy way for a user to actually get such a token and configure it in
their browser. This commit compliments another commit in
data-plane-gateway to allow users to seamlessly access HTTP resources
behind private ports by automatically fetching and configuring a
data-plane auth token for a given task.

The specific functionality introduced in this commit is entirely
contained in the `/data-plane-auth-req` page. The user is expected to
land on this page after the data-plane-gateway sees a request to a task
that is unauthorized, and redirects them here. This page expects two
parameters to be passed as part of the URL query:

- `prefix` is the authorization scope to request for the auth token
- `orig_url` is the full URL that the user is trying to access (on the
  private port)

As soon as the `/data-plane-auth-req` page loads, it will request a
data-plane auth token with the single scope given by the `prefix`
parameter. We rely on the server to reject requests for unauthorized
scopes. If the token could be acquired, then we automatically navigate
the user to a special `/auth-redirect` handler in data-plane-gateway,
providing the auth token and the `orig_url` as URL parameters.