show unauthorized error for data-plane-auth-req
If a user ends up on the `/data-plane-auth-req` page with a `prefix`
that they're not authorized to, then they will now see an error message
saying that they don't have access to the requested resource. This is a
major improvement over the current behavior in that scenario, which was
an infinite loop of redirects.

The cause of that issue was that when requesting an access token for
scopes you don't have access to, the control plane will return a
successful response with a valid token that just doesn't have any
`prefixes`. The control plane api probably ought to be fixed, but
for now I just changed `useScopedGatewayAuthToken` to throw an error
if the returned token does not contain a prefix.

Additionally, I did a bit of refactoring in that file to simplify things
based on the expectation that it should only be called with a single
prefix.