Use text instead of icons for endpoint visibility
When displaying links to endpoints exposed by connectors, use text
instead of icons to communicate the access controls around the port.
The lock icons were removed and replaced with simple "Public|Private
endpoint:" labels (with the same tooltip as before).

Additionally, the label used when displaying the endpiont in the logs
dialog (after creating or editing a task) was simlified to remove the
name of the task, since that's likely already clear from the context.