Phil/connector endpoints (#526)
* Show task endpoints after successful publication

When a capture or materialization connector exposes one or more
endpoints, we now display a link to those endpoints in the
LogDialogActions component.

The `TaskEndpoint` component was added for rendering the endpoint info.
This component displays only a single endpoint per task, since it's
designed to fit into the small space at the bottom of the logs dialog.
The plan is for a later PR to introduce another component that
shows all of the endpoints for a task. But today, no connector exposes
more than one endpoint, so it's a non-issue.

Determining the endpoint URL requires knowledge of the data-plane
hostname. In order to get the hostname, this uses the `gateway_auth`
RPC to get an auth token that's scoped to the task. This is a bit
overkill, but was done with two things in mind. One is that there's
imminent plans to implement a data-plane authentication page that
generates this token and navigates the user to a data-plane-gateway
redirect handler, and generating these scoped auth tokens will be a
requirement for that. Second is that the control plane doesn't currently
offer any other endpoint for determining the data-plane URL for a task,
so re-using gateway_auth was just expedient.

* Add data-plane authorization page

Flow tasks may expose private network ports, which can only be accessed
using a data-plane access token that authorizes the user to access the
task. This has been a difficult feature to use because we provided no
easy way for a user to actually get such a token and configure it in
their browser. This commit compliments another commit in
data-plane-gateway to allow users to seamlessly access HTTP resources
behind private ports by automatically fetching and configuring a
data-plane auth token for a given task.

The specific functionality introduced in this commit is entirely
contained in the `/data-plane-auth-req` page. The user is expected to
land on this page after the data-plane-gateway sees a request to a task
that is unauthorized, and redirects them here. This page expects two
parameters to be passed as part of the URL query:

- `prefix` is the authorization scope to request for the auth token
- `orig_url` is the full URL that the user is trying to access (on the
  private port)

As soon as the `/data-plane-auth-req` page loads, it will request a
data-plane auth token with the single scope given by the `prefix`
parameter. We rely on the server to reject requests for unauthorized
scopes. If the token could be acquired, then we automatically navigate
the user to a special `/auth-redirect` handler in data-plane-gateway,
providing the auth token and the `orig_url` as URL parameters.

* Display task endpoints in details panel

Adds a list of exposed endpoints to the details panel.
This is similar to the endpoint that's shown in the logs dialog after
you create a task, except that this component can render multiple
endpoints.

* Improve TaskEndpoint visuals

Improves the look and wording of the TaskEndpoints components. Also adds
an icon showing the visibility of the port (public vs private), with a
tooltip.

* show unauthorized error for data-plane-auth-req

If a user ends up on the `/data-plane-auth-req` page with a `prefix`
that they're not authorized to, then they will now see an error message
saying that they don't have access to the requested resource. This is a
major improvement over the current behavior in that scenario, which was
an infinite loop of redirects.

The cause of that issue was that when requesting an access token for
scopes you don't have access to, the control plane will return a
successful response with a valid token that just doesn't have any
`prefixes`. The control plane api probably ought to be fixed, but
for now I just changed `useScopedGatewayAuthToken` to throw an error
if the returned token does not contain a prefix.

Additionally, I did a bit of refactoring in that file to simplify things
based on the expectation that it should only be called with a single
prefix.

* Fix endpoint hostname validation

I had inadvertently broken this, so this just gets it right again.
Also deletes some unused debris in `en-US.ts`.

* PR: moving tasks into memo just to be safe for larger draftspecs

* PR: adding a divider to match styling and moving into memo to be safe

* Moving error into an alert box

* Use text instead of icons for endpoint visibility

When displaying links to endpoints exposed by connectors, use text
instead of icons to communicate the access controls around the port.
The lock icons were removed and replaced with simple "Public|Private
endpoint:" labels (with the same tooltip as before).

Additionally, the label used when displaying the endpiont in the logs
dialog (after creating or editing a task) was simlified to remove the
name of the task, since that's likely already clear from the context.

---------

Co-authored-by: Travis Jenkins <travis@estuary.dev>