re-word text for evolving collection schemas
Updates the text related to schema evolution to account for new behavior
of the evolutions handler. The evolutions handler will no longer create
new collections when the draft collection uses the `x-infer-schema`
attribute, unless explicitly asked to do so. So the verbiage of creating
new collections no longer seems appropriate. This tries to make it a
little better.