omit new_name when evolving collections
Updates the collection evolutions handler to no longer specify a
`new_name` when evolving collections. This is a temporary change, which
will remain in place until we implement a more comprehensive dialog for
showing the proposed changes. In the meantime, this was a convenient way
to opt in to the new behavior of the evolutions handler, which will now
skip re-creating collections that have the `x-infer-schema` attribute in
favor of only updating the materialization bindings to use a new
resource name.