Show evolution actions
This updates the handling of incompatible collections from failed publications
to account for the new behavior of the evolutions and publications handlers.
The agent's behavior around the evolution of collections has changed in a few
ways recently, and this change brings the UI in line with that.  Previously,
the evolutions handler would automatically choose between creating new
collections or just new materialization bindings, based on the presense of the
`x-infer-schema` annotation.  This was a flawed heuristic, and this commit sets
things more or less straight (at least until we identify other problems).

The UI will now determine the specific action that's required in order to
remediate a propsed change that's been deemed incompatible. It does this based
on a new field that's been added to the `IncompatibleCollection`s in the
publication `job_status`.  The new field is called `requires_recreation`, and
its value is an array of `reason`s why a collection should be re-created. If
that field is present and non-empty, then the UI will pass a `new_name` for
that collection when creating the `evolutions` job.  If `requires_recreation`
is empty, then no `new_name` will be set and the `evolutions` job will only
update materialization bindings.

The end result can be somewhat more complicated for users to reason about,
since some collections may need re-created while others may only need new
materialization bindings. The wording of various evolution-related messages has
been updated.  The UI now displays a proposed 'action' alongside each each
incompatible collection. For collections that need re-created, there's also a
reason.  Collections that _aren't_ being re-created could in theory also have a
reason, but the error messages being returned by connectors don't currently
seem appropriate for display there, so I left them out.