Phil/evolve backfill (#932)
* Update wording to match changes in evolutions handler

Updates the wording around evolution actions to reflect the new behavior of the
evolutions handler, which increments the backfill counter instead of re-naming
resources.

* Allow evolutions to work when publishing materializaions

The ResourceConfig Store had some baked in assumptions that it would be
operating only on captures. But now it's possible to trigger an evolution when
only publishing a materialization, because inferred schema changes can cause
incompatibilities. This updates the ResourceConfig Store to handle both capture
and materialization specs.

* allow auto-discover options to be set independently

This allows toggling the "Breaking changes re-version collections" and
"Automatically add new collections" properties indepently of one another. Both
of those toggles still require "Keep schemas up to date" to be enabled in order
to toggle them on. But now you're able to toggle on "Breaking changes
re-version collections" without also having to enable "Automatically add new
collections". As of this commit, you can now create a capture with:
`autoDiscover: { evolveIncompatibleCollections: true, addNewBindings: false }`

---------

Co-authored-by: Travis Jenkins <travis@estuary.dev>